refactor: Restructure diagram related modules
