+++ capellambse/aird/_visual.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 import collections.abc as cabc
 import logging
 
-from capellambse import aird
+from capellambse import diagram
 
 from . import _common as c
 from . import _edge_factories, _styling
@@ -20,7 +20,7 @@ from . import _edge_factories, _styling
 LOGGER = logging.getLogger(__name__)
 
 
-def from_xml(ebd: c.ElementBuilder) -> aird.DiagramElement:
+def from_xml(ebd: c.ElementBuilder) -> diagram.DiagramElement:
     """Deserialize a visual element."""
     el_type = ebd.data_element.attrib[c.ATT_XMT].split(":")[-1]
     if el_type not in VISUAL_TYPES:
@@ -32,7 +32,7 @@ def from_xml(ebd: c.ElementBuilder) -> aird.DiagramElement:
     return factory(ebd)
 
 
-def connector_factory(ebd: c.ElementBuilder) -> aird.Edge:
+def connector_factory(ebd: c.ElementBuilder) -> diagram.Edge:
     """Create a Connector.
 
     A Connector is an edge that connects Notes to arbitrary elements.
@@ -42,7 +42,7 @@ def connector_factory(ebd: c.ElementBuilder) -> aird.Edge:
     styleoverrides = _styling.apply_style_overrides(
         ebd.target_diagram.styleclass, styleclass, ebd.data_element
     )
-    edge = aird.Edge(
+    edge = diagram.Edge(
         bendpoints,
         uuid=ebd.data_element.attrib[c.ATT_XMID],
         styleclass=styleclass,
@@ -51,17 +51,17 @@ def connector_factory(ebd: c.ElementBuilder) -> aird.Edge:
     )
     assert edge.uuid is not None
 
-    if isinstance(target, aird.Box):
+    if isinstance(target, diagram.Box):
         _edge_factories.snaptarget(edge.points, -1, -2, target)
         target.add_context(edge.uuid)
-    if isinstance(source, aird.Box):
+    if isinstance(source, diagram.Box):
         _edge_factories.snaptarget(edge.points, 0, 1, source)
         source.add_context(edge.uuid)
 
     return edge
 
 
-def shape_factory(ebd: c.ElementBuilder) -> aird.Box:
+def shape_factory(ebd: c.ElementBuilder) -> diagram.Box:
     """Create a Shape."""
     assert ebd.target_diagram.styleclass is not None
 
@@ -81,7 +81,7 @@ def shape_factory(ebd: c.ElementBuilder) -> aird.Box:
             break
     else:
         parent = None
-        refpos = aird.Vector2D(0, 0)
+        refpos = diagram.Vector2D(0, 0)
 
     try:
         layout = next(ebd.data_element.iterchildren("layoutConstraint"))
@@ -94,7 +94,7 @@ def shape_factory(ebd: c.ElementBuilder) -> aird.Box:
         int(layout.attrib.get("x", "0")),
         int(layout.attrib.get("y", "0")),
     )
-    size = aird.Vector2D(
+    size = diagram.Vector2D(
         int(layout.attrib.get("width", "0")),
         int(layout.attrib.get("height", "0")),
     )
@@ -102,7 +102,7 @@ def shape_factory(ebd: c.ElementBuilder) -> aird.Box:
     styleoverrides = _styling.apply_visualelement_styles(
         ebd.target_diagram.styleclass, f"Box.{styleclass}", ebd.data_element
     )
-    return aird.Box(
+    return diagram.Box(
         pos,
         size,
         label=label,
@@ -115,7 +115,7 @@ def shape_factory(ebd: c.ElementBuilder) -> aird.Box:
 
 
 VISUAL_TYPES: dict[
-    str, cabc.Callable[[c.ElementBuilder], aird.DiagramElement]
+    str, cabc.Callable[[c.ElementBuilder], diagram.DiagramElement]
 ] = {
     "BasicDecorationNode": c.SkipObject.raise_,
     "Connector": connector_factory,
