+++ capellambse/diagram/__init__.py
@@ -1,0 +1,14 @@
+# SPDX-FileCopyrightText: Copyright DB Netz AG and the capellambse contributors
+# SPDX-License-Identifier: Apache-2.0
+
+"""Various diagramming related tools.
+
+This module is used to create diagrams, which can then be exported in
+various formats (such as SVG).
+"""
+# isort: off
+from ._vector2d import *
+
+from .capstyle import *
+from ._diagram import *
+from ._json_enc import *
