+++ capellambse/diagram/capstyle.py
